pyproject.toml
[project]
name = "control"
version = "0.2"
description = "manage processes"
readme = "README.txt"
authors = [
    { name = "Moritz Moeller", email = "mm@mxs.de" }
]
requires-python = ">=3.8"
dependencies = [
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "click>=8.1.8",
]

[project.urls]
Homepage = "https://github.com/mo22/control"

[project.scripts]
control = "control.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["control"]
